Time Changes
I'm scared to commit directly to master branch because I made a bunch of changes and don't want to overwrite everything.  There are still some known issues with the code.  The first instruction cannot be M, S, G20, or G21 or the program does not run correctly.  It can be F, P, G0, G1, G2, or G3.  All points must be in bounds, or the findTheta method stops working.  This means if the scale is wrong you're screwed.  Should include a message to the user that they need to move or scale their points.  Also, sometimes arcs work and sometimes they don't, no clue why.  SpeedFile is not used for anything except reading what the speeds are afterward to double check (F - the change of speed function is programmed in units/minute, speed is in units/second).  The speed will not be accurate close to the origin or on very small lines, this can't be fixed without more time precision.  Otherwise, there are some cool new features, and it works with time for optiq and line snake so we should be able to test out the motors hardware permitting.